Factor out PerElemTest
This should be useful when testing various I/O options too.
