+++ tests/Makefile.am
@@ -46,6 +46,7 @@ unit_tests_sources = \
   geom/bbox_test.C \
   geom/edge_test.C \
   geom/elem_test.C \
+  geom/elem_test.h \
   geom/node_test.C \
   geom/point_test.C \
   geom/point_test.h \
